backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
pydantic_settings
pandas
pydantic[email]
openpyxl
python-multipart
qrcode
pillow
razorpay
httpx
python-jose
motor
passlib[bcrypt]
argon2-cffi